Rename debug adapter to debug wrapper
